+++ api/src/main/java/com/flipkart/poseidon/internal/ParamValidationFilter.java
@@ -16,6 +16,7 @@
 
 package com.flipkart.poseidon.internal;
 
+import com.fasterxml.jackson.databind.JavaType;
 import com.flipkart.poseidon.api.Configuration;
 import com.flipkart.poseidon.constants.RequestConstants;
 import com.flipkart.poseidon.core.PoseidonRequest;
@@ -26,7 +27,6 @@ import com.flipkart.poseidon.model.annotations.Trace;
 import com.flipkart.poseidon.model.annotations.Version;
 import com.flipkart.poseidon.pojos.ParamPOJO;
 import com.flipkart.poseidon.pojos.ParamsPOJO;
-import com.google.common.base.Joiner;
 import flipkart.lego.api.entities.Filter;
 import flipkart.lego.api.entities.Request;
 import flipkart.lego.api.entities.Response;
@@ -39,7 +39,6 @@ import org.slf4j.Logger;
 import java.io.IOException;
 import java.util.*;
 
-import static com.flipkart.poseidon.helper.CallableNameHelper.canonicalName;
 import static org.slf4j.LoggerFactory.getLogger;
 
 @Trace(false)
@@ -102,23 +101,19 @@ public class ParamValidationFilter implements Filter {
                         // Optional param, value is not present but default is specified
                         value = defaultValue;
                     } else {
-                        value = attribute;
+                        value = parseParamValues(name, new String[] { attribute }, datatype, multivalue, param.getJavaType());
                     }
-                }
-                else if(isBodyRequest) {
+                } else if(isBodyRequest) {
                     String bodyString = poseidonRequest.getAttribute(RequestConstants.BODY);
                     if(!StringUtils.isEmpty(bodyString)) {
                         try {
-                            if ((param.getJavatype() == null || param.getJavatype().isEmpty()) &&
-                                    (param.getDatatype() == null)) {
+                            if (param.getJavaType() == null && param.getDatatype() == null) {
                                 value = bodyString;
                             } else {
-                                value = configuration.getObjectMapper().readValue(bodyString, Class.forName(param.getJavatype()));
+                                value = configuration.getObjectMapper().readValue(bodyString, param.getJavaType());
                             }
                         } catch (IOException e) {
                             logger.error("Error in reading body : {}", e.getMessage());
-                        } catch (ClassNotFoundException e) {
-                            logger.error("Error in finding class for body : {}", e.getMessage());
                         }
                     }
                     if(failOnMissingValue && value == null) {
@@ -147,7 +142,7 @@ public class ParamValidationFilter implements Filter {
                         // Optional param, value is not present but default is specified
                         value = defaultValue;
                     } else {
-                        value = parseParamValues(name, (String[]) attribute, datatype, multivalue);
+                        value = parseParamValues(name, (String[]) attribute, datatype, multivalue, param.getJavaType());
                     }
                 }
 
@@ -164,7 +159,8 @@ public class ParamValidationFilter implements Filter {
                     ParamPOJO param = pathParams.get(i);
                     String name = param.getName();
                     String internalName = param.getInternalName();
-                    Object value = null;
+                    String value = null;
+                    ParamPOJO.DataType datatype = param.getDatatype();
 
                     int pos = param.getPosition();
                     int greedyPos = param.getGreedyPosition();
@@ -204,10 +200,11 @@ public class ParamValidationFilter implements Filter {
                         throw new BadRequestException("Missing path parameter : " + name);
                     }
 
+                    Object convertedValue = parseParamValues(name, new String[] { value }, datatype, false, param.getJavaType());
                     if (internalName != null && !internalName.isEmpty()) {
-                        parsedParams.put(internalName, value);
+                        parsedParams.put(internalName, convertedValue);
                     } else {
-                        parsedParams.put(name, value);
+                        parsedParams.put(name, convertedValue);
                     }
                 }
             }
@@ -224,14 +221,14 @@ public class ParamValidationFilter implements Filter {
         return builder.toString();
     }
 
-    private Object parseParamValues(String name, String[] values, ParamPOJO.DataType datatype, boolean multivalue) throws BadRequestException {
+    private Object parseParamValues(String name, String[] values, ParamPOJO.DataType datatype, boolean multivalue, JavaType javaType) throws BadRequestException {
         try {
             if (values != null) {
                 if (!multivalue && values.length > 1) {
                     throw new BadRequestException("Multiple values provided for parameter : " + name);
                 }
 
-                List parsedValues = getValues(values, datatype);
+                List parsedValues = getValues(values, datatype, javaType);
                 return multivalue ? parsedValues : parsedValues.get(0);
             }
 
@@ -241,7 +238,7 @@ public class ParamValidationFilter implements Filter {
         }
     }
 
-    private List getValues(String[] values, ParamPOJO.DataType datatype) {
+    private List getValues(String[] values, ParamPOJO.DataType datatype, JavaType javaType) throws BadRequestException {
         switch (datatype) {
             case NUMBER:
                 return getDoubleValues(values);
@@ -249,6 +246,8 @@ public class ParamValidationFilter implements Filter {
                 return getIntegerValues(values);
             case BOOLEAN:
                 return getBooleanValues(values);
+            case ENUM:
+                return getEnumValues(values, javaType);
         }
 
         return getStringValues(values);
@@ -283,5 +282,22 @@ public class ParamValidationFilter implements Filter {
 
     private List<String> getStringValues(String[] values) {
         return Arrays.asList(values);
+    }
+
+    private List getEnumValues(String[] values, JavaType javaType) throws BadRequestException {
+        List enumValues = new ArrayList<>();
+        for (String value : values) {
+            try {
+                enumValues.add(configuration.getObjectMapper().convertValue(value, javaType));
+            } catch (IllegalArgumentException e) {
+                logger.error("Wrong value passed for enum : {}", e.getMessage());
+                throw new BadRequestException("Wrong value passed for enum, javatype: " + javaType + " value: " + value);
+            } catch (Exception e) {
+                logger.error("Error in reading enum : {}", e.getMessage());
+                throw new BadRequestException("Error in reading enum, javatype: " + javaType + " value: " + value);
+            }
+        }
+
+        return enumValues;
     }
 }