+++ api/src/test/java/com/flipkart/poseidon/internal/ParamValidationFilterTest.java
@@ -16,11 +16,18 @@
 
 package com.flipkart.poseidon.internal;
 
+import com.flipkart.poseidon.TestConfiguration;
+import com.flipkart.poseidon.TestEnum;
+import com.flipkart.poseidon.TestPOJO;
+import com.flipkart.poseidon.Wrapper;
+import com.flipkart.poseidon.api.APIBuildable;
 import com.flipkart.poseidon.api.Configuration;
 import com.flipkart.poseidon.constants.RequestConstants;
 import com.flipkart.poseidon.core.PoseidonRequest;
 import com.flipkart.poseidon.core.PoseidonResponse;
 import com.flipkart.poseidon.core.RequestContext;
+import com.flipkart.poseidon.model.VariableModel;
+import com.flipkart.poseidon.pojos.EndpointPOJO;
 import com.flipkart.poseidon.pojos.ParamPOJO;
 import com.flipkart.poseidon.pojos.ParamsPOJO;
 import flipkart.lego.api.exceptions.BadRequestException;
@@ -28,21 +35,19 @@ import org.junit.Before;
 import org.junit.Test;
 
 import javax.servlet.http.HttpServletRequest;
-
-import java.util.ArrayList;
 import java.util.Collections;
+import java.util.HashMap;
+import java.util.List;
 import java.util.Map;
 
 import static org.junit.Assert.*;
-import static org.mockito.Mockito.mock;
-import static org.mockito.Mockito.spy;
-import static org.mockito.Mockito.when;
+import static org.mockito.Mockito.*;
 
 /**
  * Created by shrey.garg on 14/07/16.
  */
 public class ParamValidationFilterTest {
-    private Configuration configuration = null;
+    private Configuration configuration = new TestConfiguration();
 
     @Before
     public void setUp() throws Exception {
@@ -258,19 +263,423 @@ public class ParamValidationFilterTest {
         assertEquals("123/567/212", parsedParams.get("test4"));
     }
 
+    @Test
+    public void testQueryParam() throws Exception {
+        ParamsPOJO params = mock(ParamsPOJO.class);
+
+        ParamPOJO paramPOJO = mockQueryParam("test", ParamPOJO.DataType.STRING);
+        ParamPOJO paramPOJO2 = mockQueryParam("test2", ParamPOJO.DataType.BOOLEAN);
+        ParamPOJO paramPOJO3 = mockQueryParam("test3", ParamPOJO.DataType.INTEGER);
+        ParamPOJO paramPOJO4 = mockQueryParam("test4", ParamPOJO.DataType.NUMBER);
+
+        ParamPOJO[] paramPOJOs = new ParamPOJO[] { paramPOJO, paramPOJO2, paramPOJO3, paramPOJO4 };
+        when(params.getRequired()).thenReturn(paramPOJOs);
+
+        Map<String, String[]> parameterMap = new HashMap<>();
+        parameterMap.put("test", new String[] { "works" });
+        parameterMap.put("test2", new String[] { "true" });
+        parameterMap.put("test3", new String[] { "2132" });
+        parameterMap.put("test4", new String[] { "32.532" });
+
+        RequestContext.set(RequestConstants.URI, "/3/abc/xyz");
+        HttpServletRequest servletRequest = mockHttpServletRequest("/3/abc/xyz", parameterMap);
+
+        PoseidonRequest request = new PoseidonRequest(servletRequest);
+        new ParamValidationFilter(params, configuration).filterRequest(request, new PoseidonResponse());
+
+        Map<String, Object> parsedParams = request.getAttribute(RequestConstants.PARAMS);
+        assertEquals(4, parsedParams.size());
+
+        Object test = parsedParams.get("test");
+        assertTrue(test instanceof String);
+        assertEquals("works", test);
+
+        Object test2 = parsedParams.get("test2");
+        assertTrue(test2 instanceof Boolean);
+        assertEquals(true, test2);
+
+        Object test3 = parsedParams.get("test3");
+        assertTrue(test3 instanceof Integer);
+        assertEquals(2132, test3);
+
+        Object test4 = parsedParams.get("test4");
+        assertTrue(test4 instanceof Double);
+        assertEquals(32.532, test4);
+    }
+
+    @Test
+    public void testHeaderParam() throws Exception {
+        ParamsPOJO params = mock(ParamsPOJO.class);
+
+        ParamPOJO paramPOJO = mockHeaderParam("test", ParamPOJO.DataType.STRING);
+        ParamPOJO paramPOJO2 = mockHeaderParam("test2", ParamPOJO.DataType.BOOLEAN);
+        ParamPOJO paramPOJO3 = mockHeaderParam("test3", ParamPOJO.DataType.INTEGER);
+        ParamPOJO paramPOJO4 = mockHeaderParam("test4", ParamPOJO.DataType.NUMBER);
+
+        ParamPOJO[] paramPOJOs = new ParamPOJO[] { paramPOJO, paramPOJO2, paramPOJO3, paramPOJO4 };
+        when(params.getRequired()).thenReturn(paramPOJOs);
+
+        Map<String, String> headers = new HashMap<>();
+        headers.put("test", "works");
+        headers.put("test2", "true");
+        headers.put("test3", "2132");
+        headers.put("test4", "32.532");
+
+        RequestContext.set(RequestConstants.URI, "/3/abc/xyz");
+        HttpServletRequest servletRequest = mockHttpServletRequest("/3/abc/xyz", new HashMap<>(), headers);
+
+        PoseidonRequest request = new PoseidonRequest(servletRequest);
+        new ParamValidationFilter(params, configuration).filterRequest(request, new PoseidonResponse());
+
+        Map<String, Object> parsedParams = request.getAttribute(RequestConstants.PARAMS);
+        assertEquals(4, parsedParams.size());
+
+        Object test = parsedParams.get("test");
+        assertTrue(test instanceof String);
+        assertEquals("works", test);
+
+        Object test2 = parsedParams.get("test2");
+        assertTrue(test2 instanceof Boolean);
+        assertEquals(true, test2);
+
+        Object test3 = parsedParams.get("test3");
+        assertTrue(test3 instanceof Integer);
+        assertEquals(2132, test3);
+
+        Object test4 = parsedParams.get("test4");
+        assertTrue(test4 instanceof Double);
+        assertEquals(32.532, test4);
+    }
+
+    @Test
+    public void testPathParamTypes() throws Exception {
+        ParamsPOJO params = mock(ParamsPOJO.class);
+
+        ParamPOJO paramPOJO = mockPathParam("test", 3, ParamPOJO.DataType.STRING);
+        ParamPOJO paramPOJO2 = mockPathParam("test2", 4, ParamPOJO.DataType.BOOLEAN);
+        ParamPOJO paramPOJO3 = mockPathParam("test3", 5, ParamPOJO.DataType.INTEGER);
+        ParamPOJO paramPOJO4 = mockPathParam("test4", 6, ParamPOJO.DataType.NUMBER);
+
+        ParamPOJO[] paramPOJOs = new ParamPOJO[] { paramPOJO, paramPOJO2, paramPOJO3, paramPOJO4 };
+        when(params.getRequired()).thenReturn(paramPOJOs);
+
+        RequestContext.set(RequestConstants.URI, "/3/abc/{test}/{test2}/{test3}/{test4}");
+        HttpServletRequest servletRequest = mockHttpServletRequest("/3/abc/works/true/2132/32.532");
+
+        PoseidonRequest request = new PoseidonRequest(servletRequest);
+        new ParamValidationFilter(params, configuration).filterRequest(request, new PoseidonResponse());
+
+        Map<String, Object> parsedParams = request.getAttribute(RequestConstants.PARAMS);
+        assertEquals(4, parsedParams.size());
+
+        Object test = parsedParams.get("test");
+        assertTrue(test instanceof String);
+        assertEquals("works", test);
+
+        Object test2 = parsedParams.get("test2");
+        assertTrue(test2 instanceof Boolean);
+        assertEquals(true, test2);
+
+        Object test3 = parsedParams.get("test3");
+        assertTrue(test3 instanceof Integer);
+        assertEquals(2132, test3);
+
+        Object test4 = parsedParams.get("test4");
+        assertTrue(test4 instanceof Double);
+        assertEquals(32.532, test4);
+    }
+
+    @Test
+    public void testEnumParams() throws Exception {
+        ParamsPOJO params = mock(ParamsPOJO.class);
+
+        ParamPOJO paramPOJO = mockQueryParam("test", ParamPOJO.DataType.ENUM);
+        mockJavaType(paramPOJO, "com.flipkart.poseidon.TestEnum");
+        ParamPOJO paramPOJO1 = mockHeaderParam("test1", ParamPOJO.DataType.ENUM);
+        mockJavaType(paramPOJO1, "com.flipkart.poseidon.TestEnum");
+        ParamPOJO paramPOJO2 = mockPathParam("test2", 2, ParamPOJO.DataType.ENUM);
+        mockJavaType(paramPOJO2, "com.flipkart.poseidon.TestEnum");
+
+        ParamPOJO[] paramPOJOs = new ParamPOJO[] { paramPOJO, paramPOJO1, paramPOJO2 };
+        when(params.getRequired()).thenReturn(paramPOJOs);
+        load(params);
+
+        Map<String, String[]> parameterMap = new HashMap<>();
+        parameterMap.put("test", new String[] { "XYZ" });
+
+        Map<String, String> headers = new HashMap<>();
+        headers.put("test1", "ABC");
+
+        RequestContext.set(RequestConstants.URI, "/3/{test2}/xyz");
+        HttpServletRequest servletRequest = mockHttpServletRequest("/3/ABC/xyz", parameterMap, headers);
+
+        PoseidonRequest request = new PoseidonRequest(servletRequest);
+        new ParamValidationFilter(params, configuration).filterRequest(request, new PoseidonResponse());
+
+        Map<String, Object> parsedParams = request.getAttribute(RequestConstants.PARAMS);
+        assertEquals(3, parsedParams.size());
+
+        Object test = parsedParams.get("test");
+        assertTrue(test instanceof TestEnum);
+        assertEquals(TestEnum.XYZ, test);
+
+        Object testPath = parsedParams.get("test2");
+        assertTrue(testPath instanceof TestEnum);
+        assertEquals(TestEnum.ABC, testPath);
+
+        Object testHeader = parsedParams.get("test1");
+        assertTrue(testHeader instanceof TestEnum);
+        assertEquals(TestEnum.ABC, testHeader);
+    }
+
+    @Test(expected = BadRequestException.class)
+    public void testQueryParamEnumWrong() throws Exception {
+        ParamsPOJO params = mock(ParamsPOJO.class);
+
+        ParamPOJO paramPOJO = mockQueryParam("test", ParamPOJO.DataType.ENUM);
+        mockJavaType(paramPOJO, "com.flipkart.poseidon.TestEnum");
+
+        ParamPOJO[] paramPOJOs = new ParamPOJO[] { paramPOJO };
+        when(params.getRequired()).thenReturn(paramPOJOs);
+
+        Map<String, String[]> parameterMap = new HashMap<>();
+        parameterMap.put("test", new String[] { "XYZ-A" });
+
+        RequestContext.set(RequestConstants.URI, "/3/abc/xyz");
+        HttpServletRequest servletRequest = mockHttpServletRequest("/3/abc/xyz", parameterMap);
+
+        PoseidonRequest request = new PoseidonRequest(servletRequest);
+        new ParamValidationFilter(params, configuration).filterRequest(request, new PoseidonResponse());
+        fail();
+    }
+
+    @Test(expected = UnsupportedOperationException.class)
+    public void testQueryParamEnumWrongClass() throws Exception {
+        ParamsPOJO params = mock(ParamsPOJO.class);
+
+        ParamPOJO paramPOJO = mockQueryParam("test", ParamPOJO.DataType.ENUM);
+        mockJavaType(paramPOJO, "com.flipkart.poseidon.TestEnum123");
+
+        ParamPOJO[] paramPOJOs = new ParamPOJO[] { paramPOJO };
+        when(params.getRequired()).thenReturn(paramPOJOs);
+        load(params);
+
+        Map<String, String[]> parameterMap = new HashMap<>();
+        parameterMap.put("test", new String[] { "XYZ-A" });
+
+        RequestContext.set(RequestConstants.URI, "/3/abc/xyz");
+        HttpServletRequest servletRequest = mockHttpServletRequest("/3/abc/xyz", parameterMap);
+
+        PoseidonRequest request = new PoseidonRequest(servletRequest);
+        new ParamValidationFilter(params, configuration).filterRequest(request, new PoseidonResponse());
+        fail();
+    }
+
+    @Test
+    public void testBodyParam() throws Exception {
+        ParamsPOJO params = mock(ParamsPOJO.class);
+
+        ParamPOJO paramPOJO = mockBodyParam("test", "com.flipkart.poseidon.TestPOJO");
+
+        ParamPOJO[] paramPOJOs = new ParamPOJO[] { paramPOJO };
+        when(params.getRequired()).thenReturn(paramPOJOs);
+        load(params);
+
+        RequestContext.set(RequestConstants.URI, "/3/abc/xyz");
+        HttpServletRequest servletRequest = mockHttpServletRequest("/3/abc/xyz");
+
+        PoseidonRequest request = new PoseidonRequest(servletRequest);
+        request.setAttribute(RequestConstants.BODY, configuration.getObjectMapper().writeValueAsString(new TestPOJO("xyz", true)));
+
+        new ParamValidationFilter(params, configuration).filterRequest(request, new PoseidonResponse());
+
+        Map<String, Object> parsedParams = request.getAttribute(RequestConstants.PARAMS);
+        assertEquals(1, parsedParams.size());
+
+        Object test = parsedParams.get("test");
+        assertTrue(test instanceof TestPOJO);
+        TestPOJO testPOJO = (TestPOJO) test;
+        assertEquals("xyz", testPOJO.getAbc());
+        assertEquals(true, testPOJO.isTest());
+    }
+
+    @Test
+    public void testBodyParamList() throws Exception {
+        ParamsPOJO params = mock(ParamsPOJO.class);
+
+        ParamPOJO paramPOJO = mockBodyParam("test",
+                new VariableModel("java.util.List",
+                        new VariableModel[] { new VariableModel("com.flipkart.poseidon.TestPOJO") }));
+
+        ParamPOJO[] paramPOJOs = new ParamPOJO[] { paramPOJO };
+        when(params.getRequired()).thenReturn(paramPOJOs);
+        load(params);
+
+        RequestContext.set(RequestConstants.URI, "/3/abc/xyz");
+        HttpServletRequest servletRequest = mockHttpServletRequest("/3/abc/xyz");
+
+        PoseidonRequest request = new PoseidonRequest(servletRequest);
+        request.setAttribute(RequestConstants.BODY, configuration.getObjectMapper().writeValueAsString(Collections.singletonList(new TestPOJO("xyz", true))));
+
+        new ParamValidationFilter(params, configuration).filterRequest(request, new PoseidonResponse());
+
+        Map<String, Object> parsedParams = request.getAttribute(RequestConstants.PARAMS);
+        assertEquals(1, parsedParams.size());
+
+        Object test = parsedParams.get("test");
+        assertTrue(test instanceof List);
+        Object testPOJOBlob = ((List) test).get(0);
+        assertTrue(testPOJOBlob instanceof TestPOJO);
+        TestPOJO testPOJO = (TestPOJO) testPOJOBlob;
+        assertEquals("xyz", testPOJO.getAbc());
+        assertEquals(true, testPOJO.isTest());
+    }
+
+    @Test
+    public void testBodyParamMap() throws Exception {
+        ParamsPOJO params = mock(ParamsPOJO.class);
+
+        ParamPOJO paramPOJO = mockBodyParam("test",
+                new VariableModel("java.util.Map", new VariableModel[] {
+                        new VariableModel("java.lang.String"),
+                        new VariableModel("com.flipkart.poseidon.TestPOJO")
+                }));
+
+        ParamPOJO[] paramPOJOs = new ParamPOJO[] { paramPOJO };
+        when(params.getRequired()).thenReturn(paramPOJOs);
+        load(params);
+
+        RequestContext.set(RequestConstants.URI, "/3/abc/xyz");
+        HttpServletRequest servletRequest = mockHttpServletRequest("/3/abc/xyz");
+
+        PoseidonRequest request = new PoseidonRequest(servletRequest);
+        request.setAttribute(RequestConstants.BODY, configuration.getObjectMapper().writeValueAsString(Collections.singletonMap("key", new TestPOJO("xyz", true))));
+
+        new ParamValidationFilter(params, configuration).filterRequest(request, new PoseidonResponse());
+
+        Map<String, Object> parsedParams = request.getAttribute(RequestConstants.PARAMS);
+        assertEquals(1, parsedParams.size());
+
+        Object test = parsedParams.get("test");
+        assertTrue(test instanceof Map);
+        Object testPOJOBlob = ((Map) test).get("key");
+        assertTrue(testPOJOBlob instanceof TestPOJO);
+        TestPOJO testPOJO = (TestPOJO) testPOJOBlob;
+        assertEquals("xyz", testPOJO.getAbc());
+        assertEquals(true, testPOJO.isTest());
+    }
+
+    @Test
+    public void testBodyParamGenerics() throws Exception {
+        ParamsPOJO params = mock(ParamsPOJO.class);
+
+        ParamPOJO paramPOJO = mockBodyParam("test",
+                new VariableModel("com.flipkart.poseidon.Wrapper", new VariableModel[] {
+                        new VariableModel("java.util.Map", new VariableModel[] {
+                                new VariableModel("java.lang.String"),
+                                new VariableModel("com.flipkart.poseidon.TestPOJO")
+                        })
+                }));
+
+        ParamPOJO[] paramPOJOs = new ParamPOJO[] { paramPOJO };
+        when(params.getRequired()).thenReturn(paramPOJOs);
+        load(params);
+
+        RequestContext.set(RequestConstants.URI, "/3/abc/xyz");
+        HttpServletRequest servletRequest = mockHttpServletRequest("/3/abc/xyz");
+
+        PoseidonRequest request = new PoseidonRequest(servletRequest);
+        request.setAttribute(RequestConstants.BODY, configuration.getObjectMapper().writeValueAsString(new Wrapper<>(Collections.singletonMap("key", new TestPOJO("xyz", true)))));
+
+        new ParamValidationFilter(params, configuration).filterRequest(request, new PoseidonResponse());
+
+        Map<String, Object> parsedParams = request.getAttribute(RequestConstants.PARAMS);
+        assertEquals(1, parsedParams.size());
+
+        Object test = parsedParams.get("test");
+        assertTrue(test instanceof Wrapper);
+        Object wrapperMap = ((Wrapper) test).getObj();
+        assertTrue(wrapperMap instanceof Map);
+        Object testPOJOBlob = ((Map) wrapperMap).get("key");
+        assertTrue(testPOJOBlob instanceof TestPOJO);
+        TestPOJO testPOJO = (TestPOJO) testPOJOBlob;
+        assertEquals("xyz", testPOJO.getAbc());
+        assertEquals(true, testPOJO.isTest());
+    }
+
     private ParamPOJO mockPathParam(String name, int position) {
+        return mockPathParam(name, position, ParamPOJO.DataType.STRING);
+    }
+
+    private ParamPOJO mockPathParam(String name, int position, ParamPOJO.DataType type) {
         ParamPOJO paramPOJO = spy(ParamPOJO.class);
         when(paramPOJO.getName()).thenReturn(name);
         when(paramPOJO.getDatatype()).thenReturn(ParamPOJO.DataType.STRING);
         when(paramPOJO.isPathparam()).thenReturn(true);
         when(paramPOJO.getPosition()).thenReturn(position);
+        when(paramPOJO.getDatatype()).thenReturn(type);
         return paramPOJO;
     }
 
+    private ParamPOJO mockQueryParam(String name, ParamPOJO.DataType type) {
+        ParamPOJO paramPOJO = spy(ParamPOJO.class);
+        when(paramPOJO.getName()).thenReturn(name);
+        when(paramPOJO.getDatatype()).thenReturn(type);
+        return paramPOJO;
+    }
+
+    private ParamPOJO mockBodyParam(String name, String javaType) {
+        ParamPOJO paramPOJO = spy(ParamPOJO.class);
+        when(paramPOJO.getName()).thenReturn(name);
+        when(paramPOJO.getType()).thenReturn(new VariableModel(javaType));
+        when(paramPOJO.getJavatype()).thenReturn(javaType);
+        when(paramPOJO.isBody()).thenReturn(true);
+        return paramPOJO;
+    }
+
+    private ParamPOJO mockBodyParam(String name, VariableModel variableModel) {
+        ParamPOJO paramPOJO = spy(ParamPOJO.class);
+        when(paramPOJO.getName()).thenReturn(name);
+        when(paramPOJO.getType()).thenReturn(variableModel);
+        when(paramPOJO.isBody()).thenReturn(true);
+        return paramPOJO;
+    }
+
+    private ParamPOJO mockHeaderParam(String name, ParamPOJO.DataType type) {
+        ParamPOJO paramPOJO = spy(ParamPOJO.class);
+        when(paramPOJO.getName()).thenReturn(name);
+        when(paramPOJO.getDatatype()).thenReturn(type);
+        when(paramPOJO.isHeader()).thenReturn(true);
+        return paramPOJO;
+    }
+
+    private void mockJavaType(ParamPOJO paramPOJO, String javaType) {
+        when(paramPOJO.getJavatype()).thenReturn(javaType);
+    }
+
     private HttpServletRequest mockHttpServletRequest(String path) {
+        return mockHttpServletRequest(path, new HashMap<>(), new HashMap<>());
+    }
+
+    private HttpServletRequest mockHttpServletRequest(String path, Map<String, String[]> parameterMap) {
+        return mockHttpServletRequest(path, parameterMap, new HashMap<>());
+    }
+
+    private HttpServletRequest mockHttpServletRequest(String path, Map<String, String[]> parameterMap, Map<String, String> headers) {
         HttpServletRequest servletRequest = mock(HttpServletRequest.class);
         when(servletRequest.getPathInfo()).thenReturn(path);
-        when(servletRequest.getHeaderNames()).thenReturn(Collections.enumeration(new ArrayList<>()));
+        when(servletRequest.getHeaderNames()).thenReturn(Collections.enumeration(headers.keySet()));
+        headers.forEach((h, v) -> {
+            when(servletRequest.getHeader(eq(h))).thenReturn(v);
+        });
+        when(servletRequest.getParameterMap()).thenReturn(parameterMap);
         return servletRequest;
+    }
+
+    private void load(ParamsPOJO params) {
+        EndpointPOJO endpointPOJO = spy(EndpointPOJO.class);
+        when(endpointPOJO.getParams()).thenReturn(params);
+
+        new APIBuildable(null, endpointPOJO, configuration, new HashMap<>());
     }
 }