+++ cadfael/src/main/java/com/flipkart/poseidon/cadfael/artifact/Scope.java
@@ -1,0 +1,12 @@
+package com.flipkart.poseidon.cadfael.artifact;
+
+/**
+ * Created by shrey.garg on 02/09/17.
+ */
+public enum Scope {
+    provided,
+    test,
+    compile,
+    runtime,
+    system
+}